Add api, common subprojects
