+++ api/build.sbt
@@ -1,0 +1,5 @@
+import sbt.Keys.name
+
+name := "grafink-api"
+
+enablePlugins(BuildInfoPlugin, JavaServerAppPackaging)
