+++ api/src/main/scala/com/astrolabsoftware/grafink/Boot.scala
@@ -1,0 +1,3 @@
+package com.astrolabsoftware.grafink
+
+object Boot {}
