+++ build.sbt
@@ -13,6 +13,22 @@ lazy val compileScalastyle = taskKey[Unit]("compileScalastyle")
 compileScalastyle := scalastyle.in(Compile).toTask("").value
 (compile in Compile) := ((compile in Compile) dependsOn compileScalastyle).value
 
+lazy val api =
+  (project in file("api"))
+    .settings(
+      // Add support for scala version 2.11
+      crossScalaVersions := Seq("2.11.11", (ThisBuild / scalaVersion).value),
+      apiSettings
+    ).dependsOn(common)
+
+lazy val common =
+  (project in file("common"))
+    .settings(
+      // Add support for scala version 2.11
+      crossScalaVersions := Seq("2.11.11", (ThisBuild / scalaVersion).value),
+      commonSettings
+    )
+
 lazy val core =
   (project in file("core"))
     .settings(
@@ -21,17 +37,19 @@ lazy val core =
       buildInfoKeys := Seq[BuildInfoKey](name, version, scalaVersion, sbtVersion),
       buildInfoPackage := "com.astrolabsoftware.grafink",
       stdSettings
-    )
+    ).dependsOn(common)
 
 lazy val js =
   (project in file("js"))
     .settings(
-      commonSettings ++ scalaJSSettings
+      coverageEnabled := false,
+      resolverSettings ++ scalaJSSettings
     ).enablePlugins(ScalaJSPlugin, ScalaJSBundlerPlugin)
 
 lazy val root = (project in file(".")).settings(
   skip in publish := true
 ).aggregate(
+  api,
   core,
   js
 )
