+++ common/src/main/scala/com/astrolabsoftware/grafink/logging/Logger.scala
