+++ common/src/main/scala/com/astrolabsoftware/grafink/models/GrafinkException.scala
