+++ core/src/main/scala/com/astrolabsoftware/grafink/common/Utils.scala
@@ -61,15 +61,4 @@ object Utils {
           .fold(failure => log.error(s"Error closing filesystem: $failure"), _ => ZIO.succeed(Unit))
       )(fs => f(fs))
 
-  /**
-   * Split a string into tuple, based on first match
-   * @param s
-   */
-  implicit class StringToTuple(s: String) {
-    def splitToTuple(regex: String): (String, String) =
-      s.split(regex, 2) match {
-        case Array(str1, str2) => (str1, str2)
-        case Array(str1)       => (str1, "")
-      }
-  }
 }
