+++ js/src/main/scala/com/astrolabsoftware/grafink/GrafinkApp.scala
@@ -16,24 +16,9 @@ object GrafinkApp {
     targetNode.appendChild(parNode)
   }
 
-  def addClickedMessage(): Unit = {
-    appendPar(document.body, "You clicked the button!")
-  }
+  def setupUI(): Unit =
+    appendPar(document.body, "Hello Grafink")
 
-  def setupUI(): Unit = {
-    val button = document.createElement("button")
-    button.textContent = "Click me!"
-    button.addEventListener("click", { (e: dom.MouseEvent) =>
-      addClickedMessage()
-    })
-    document.body.appendChild(button)
-
-    appendPar(document.body, "Hello World")
-  }
-
-  def main(args: Array[String]): Unit = {
-    document.addEventListener("DOMContentLoaded", { (e: dom.Event) =>
-      setupUI()
-    })
-  }
+  def main(args: Array[String]): Unit =
+    document.addEventListener("DOMContentLoaded", (e: dom.Event) => setupUI())
 }
