+++ project/plugins.sbt
@@ -14,3 +14,5 @@ addSbtPlugin("org.scala-js" % "sbt-scalajs" % "1.1.1")
 
 addSbtPlugin("ch.epfl.scala" % "sbt-scalajs-bundler" % "0.18.0")
 
+libraryDependencies += "org.scala-js" %% "scalajs-env-jsdom-nodejs" % "1.0.0"
+
